+++ _h/crc8.c
@@ -10,5 +10,5 @@ uint8_t CRC_bytes( uint8_t* data, uint8_t len ) {
 }
 
 uint8_t CRC_byte( uint8_t oldCRC, uint8_t data ) {
-    return pgm_read_byte( CRC_table + ((uint8_t)( oldCRC << 1 ) ^ data) );
+    return pgm_read_byte( CRC_table + (uint8_t)((uint8_t)( oldCRC << 1 ) ^ data) );
 }
